+++ pystan/plots.py
@@ -4,16 +4,13 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger('pystan')
 
 
-def traceplot(fit, vars=None):
+def traceplot(fit, pars):
     """Use pymc's traceplot to display parameters"""
     # FIXME: eventually put this in the StanFit object
     # FIXME: write a to_pymc(_trace) function
-    samples = fit.extract()
-    if vars is None:
-        vars = [v for v in samples.keys() if v != 'lp__']
     try:
         from pystan.external.pymc import plots
     except ImportError:
         logger.critical("matplotlib required for plotting.")
         raise
-    return plots.traceplot(fit.extract(), vars)
+    return plots.traceplot(fit.extract(), pars)